requirements.txt
amqp==2.3.2
asgiref==3.2.3
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
celery-once==2.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
configparser==4.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.1.3
django-celery-beat==1.3.0
django-celery-results==1.0.4
django-timezone-field==3.0
djangorestframework==3.9.0
dukpy==0.2.2
echarts-themes-pypkg==0.0.3
future==0.17.1
idna==2.7
imgkit==1.0.1
importlib-metadata==1.5.0
itypes==1.1.0
javascripthon==0.10
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
jwt==0.5.4
kombu==4.2.1
lml==0.0.2
macropy3==1.1.0b2
MarkupSafe==1.1.0
mysqlclient==1.3.13
numpy==1.16.1
pandas==1.0.1
Pillow==5.4.1
pycparser==2.19
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.10
PyMySQL==0.9.2
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.7
redis==2.10.6
requests==2.20.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
zipp==3.1.0
xlwt==1.3.0


